[oo] new Principal.java class and gitignore file are defined
