+++ oo/.gitignore
@@ -1,0 +1,4 @@
+/.metadata/
+/.recommenders/
+IRPF/.classpath
+IRPF/.project
